tests: move mocks
